backend/pyproject.toml
[project]
name = "codebase-cartographer-backend"
version = "0.1.0"
description = "Structural forensics backend for Codebase Cartographer"
requires-python = ">=3.11,<3.15"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "neo4j>=5.23.0",
  "httpx>=0.27.0",
  "GitPython>=3.1.43",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "bandit>=1.7.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S"]
ignore = ["S101"]

